# fragtale-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fragtale_api",
    "description": "Fragtale API",
    "license": {
      "name": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/topics/{topic_id}/confirm": {
      "get": {
        "tags": [
          "web_socket"
        ],
        "summary": "Open a WebSocket connection for sending event delivery confirmation messages.",
        "description": "Consumer identifier is derived from authentication.",
        "operationId": "confirm_event_delivery",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching protocols to websocket."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/topics/{topic_id}/confirm/{unique_time}/{instance_id}": {
      "put": {
        "tags": [
          "http"
        ],
        "summary": "Confirm successful delivery of an event.",
        "description": "Consumer identifier is derived from authentication.",
        "operationId": "confirm_event_delivery",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unique_time",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully confirmed event delivery."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          },
          "403": {
            "description": "Forbidden: Authorization failure."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/topics/{topic_id}/correlation/{correlation_token}": {
      "get": {
        "tags": [
          "http"
        ],
        "summary": "Retrieve event document related to another event.",
        "description": "This enables treating the API as an RPC-style REST API with a PRG twist,\npubishing an event can be seen as a request and following the link will\nreturn an event from a different topic as a result as long as this happens\nwithin a reasonable timeout.\n\nConsumer identifier is derived from authentication.",
        "operationId": "by_topic_and_correlation_token",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "correlation_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event document is delivered in the response body.",
            "content": {
              "application/json": {}
            }
          },
          "204": {
            "description": "No correlated event document has appeared within the timeout."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          },
          "403": {
            "description": "Forbidden: Authorization failure."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/topics/{topic_id}/events/by_event_id/{event_id}": {
      "get": {
        "tags": [
          "http"
        ],
        "summary": "Retrieve an event document by its identifier.",
        "description": "Consumer identifier is derived from authentication.",
        "operationId": "event_by_topic_and_id",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "index_name",
            "in": "path",
            "description": "The name of the index.",
            "required": true
          },
          {
            "name": "index_key",
            "in": "path",
            "description": "The lookup key to use when searching the index.",
            "required": true
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the event document.",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          },
          "403": {
            "description": "Forbidden: Authorization failure."
          },
          "404": {
            "description": "No event document with the event identifier was found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/topics/{topic_id}/next": {
      "get": {
        "tags": [
          "http"
        ],
        "summary": "Poll for new events.",
        "description": "Consumer identifier is derived from authentication.",
        "operationId": "next_event_by_topic_and_consumer",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Only consider events newer than this in epoch milliseconds.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "version",
            "in": "query",
            "description": "Event Descriptor SemVer that the client prefers (major.minor).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A new event is delivered in the response body.",
            "links": {
              "Location": {
                "operation_id": "confirm_event_delivery",
                "parameters": {
                  "correlation_token": "string",
                  "instance_id": "number",
                  "topic_id": "string",
                  "unique_time": "number"
                }
              }
            }
          },
          "204": {
            "description": "No new event was found."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          },
          "403": {
            "description": "Forbidden: Authorization failure."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}